+++ pages/services.js
@@ -1,0 +1,5 @@
+import WhatWeOffer from '@/components/services/WhatWeOffer';
+
+export default function Services() {
+  return <WhatWeOffer />;
+}